@@ -20,52 +20,53 @@
   },
   "main": "dist/vue-avatar.min.js",
   "dependencies": {
-    "vue": "^2.0.0",
-    "babel-runtime": "^5.8.0"
+    "babel-runtime": "^6.26.0",
+    "vue": "^2.5.2"
   },
   "devDependencies": {
-    "babel-core": "^6.0.0",
-    "babel-loader": "^6.0.0",
-    "babel-plugin-transform-runtime": "^6.0.0",
-    "babel-preset-es2015": "^6.0.0",
-    "babel-preset-stage-2": "^6.0.0",
-    "copy-webpack-plugin": "^1.1.1",
-    "cross-env": "^1.0.6",
-    "css-loader": "^0.23.0",
-    "eslint": "^2.0.0",
-    "eslint-config-standard": "^5.1.0",
-    "eslint-friendly-formatter": "^1.2.2",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-html": "^1.3.0",
-    "eslint-plugin-promise": "^1.0.8",
-    "eslint-plugin-standard": "^1.3.2",
-    "file-loader": "^0.8.4",
-    "json-loader": "^0.5.4",
-    "url-loader": "^0.5.7",
-    "vue-hot-reload-api": "^1.2.0",
-    "vue-html-loader": "^1.0.0",
-    "vue-loader": "^8.2.1",
-    "vue-style-loader": "^1.0.0",
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.12.0",
-    "webpack-merge": "^0.8.3",
-
-    "function-bind": "^1.0.2",
-    "karma": "^0.13.15",
-    "karma-coverage": "^0.5.5",
-    "karma-mocha": "^0.2.2",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "karma-sinon-chai": "^1.2.0",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.24",
-    "karma-webpack": "^1.7.0",
-    "lolex": "^1.4.0",
-    "mocha": "^2.4.5",
-    "chai": "^3.5.0",
-    "sinon": "^1.17.3",
-    "sinon-chai": "^2.8.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.6.1",
+    "chai": "^4.1.2",
+    "copy-webpack-plugin": "^4.2.0",
+    "cross-env": "^5.1.0",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.9.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-friendly-formatter": "^3.0.0",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-html": "^3.2.2",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-node": "^5.2.0",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "file-loader": "^1.1.5",
+    "function-bind": "^1.1.1",
     "isparta-loader": "^2.0.0",
-    "phantomjs-prebuilt": "^2.1.3",
-    "ora": "^0.2.0"
+    "json-loader": "^0.5.7",
+    "karma": "^1.7.1",
+    "karma-coverage": "^1.1.1",
+    "karma-mocha": "^1.3.0",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "karma-sinon-chai": "^1.3.2",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-spec-reporter": "^0.0.31",
+    "karma-webpack": "^2.0.5",
+    "lolex": "^2.1.3",
+    "mocha": "^4.0.1",
+    "ora": "^1.3.0",
+    "phantomjs-prebuilt": "^2.1.15",
+    "sinon": "^4.0.1",
+    "sinon-chai": "^2.14.0",
+    "url-loader": "^0.6.2",
+    "vue-hot-reload-api": "^2.2.0",
+    "vue-html-loader": "^1.2.4",
+    "vue-loader": "^13.3.0",
+    "vue-style-loader": "^3.0.3",
+    "vue-template-compiler": "^2.5.2",
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.3",
+    "webpack-merge": "^4.1.0"
   }
 }
